fix: missing import + filters not loading
